+++ src/logging_strict/__init__.py
@@ -67,7 +67,7 @@ e.g. ``type[LoggingYamlType]``
 Exceptions
 -----------
 
-:ref:`[exc docs] <code/constants/exceptions:public api>`
+:ref:`[exc docs] <code/constants/exceptions:exceptions>`
 
 - LoggingStrictError